# Tweak Nolan's PDF/CDF calculations for improved accuracy and performance