I made a few tweaks (in order of decreasing importance for accuracy):

1. Add bisection-based sampling of the integrand tails in order to explicitly provide QUADPACK the function support and detection of rapidly descending tail behavior. This essentially fixes the issue with "long zero tails" noted in a previous comment.
2. Implement Nolan's rounding of difficult inputs in accordance with the information from his STABLE program. This mostly helps with `x0` being very close to `zeta` and `alpha` being very close (but not equal) to 1.0.
3. Slightly re-express Nolan's g function to improve numerical stability
4. Eliminate dependence on `np.complex128` for exponentiation on `(x0 - zeta)`
5. Refactor special functions from Nolan's method to reduce code repetition and improve performance. This also makes the code logic easier to follow (in my opinion).

This resolves most of the remaining numerical issues that I know of. See below

![bsdz_piecewise_proposal_accuracies](https://user-images.githubusercontent.com/14023456/91498191-560dd000-e88d-11ea-8a56-a5e4c1387f86.png)

Importantly, there is virtually no performance impact here. Here's a quick run on a 11x7x7 grid near the distribution mode.

```
# BEFORE TWEAKS

$ python3 -m timeit -s "from scipy.stats import levy_stable; import numpy as np" "[levy_stable.cdf(x, alpha=alpha, beta=beta) for x in np.linspace(-10, 10, 11) for alpha in np.linspace(0.5, 2.0, 7) for beta in np.linspace(0.0, 1.0, 7)]"
10 loops, best of 3: 788 msec per loop
$ python3 -m timeit -s "from scipy.stats import levy_stable; import numpy as np" "[levy_stable.pdf(x, alpha=alpha, beta=beta) for x in np.linspace(-10, 10, 11) for alpha in np.linspace(0.5, 2.0, 7) for beta in np.linspace(0.0, 1.0, 7)]"
10 loops, best of 3: 1.06 sec per loop

# AFTER TWEAKS

$ python3 -m timeit -s "from scipy.stats import levy_stable; import numpy as np" "[levy_stable.cdf(x, alpha=alpha, beta=beta) for x in np.linspace(-10, 10, 11) for alpha in np.linspace(0.5, 2.0, 7) for beta in np.linspace(0.0, 1.0, 7)]"
10 loops, best of 3: 939 msec per loop
$ python3 -m timeit -s "from scipy.stats import levy_stable; import numpy as np" "[levy_stable.pdf(x, alpha=alpha, beta=beta) for x in np.linspace(-10, 10, 11) for alpha in np.linspace(0.5, 2.0, 7) for beta in np.linspace(0.0, 1.0, 7)]"
10 loops, best of 3: 951 msec per loop
```