@ragibson 2020-09-01
This was merged manually in https://github.com/scipy/scipy/pull/9523/commits/1c458aa61bb7dae3cae0e562fdd2c7d774217dfa.